Created shotgun model and animations and integrate this into game. Currently, the Shotgun's model has some issues displaying properly (the player must be selected and viewed (selecting and holding "f") within the scene view in order for semi-consistent display.
Changed the way in which the "fire" method of the "shoot" class works, incorporating an optional delay for weapons between mouse click and shooting (used in order for animations to sync with shooting correctly).

Changed the way in which the user joins the game: created buttons and an input text field, replacing the antiquated GUI used previously. In order to incorporate this, changes were made throughout the gamemanager and network manager classes allowing for use of button clicking.

Not sure how well this commit will turn out; not sure if I managed to ensure that the builds are not committed.