Added a feature wherein the "Alive" enemies will attack you despite their "alert" not being set off (is set off when an "Alive" is killed) if the player stays too close to the enemy for too long.
Fixed a bug where the shotgun was not visible initially.

Added a simple texture to the shotgun: it is now black and semi-transparent throughout most of the body.
Changed the way in which the game ends to be when all of the "Static" enemies are killed, rather than both the "Alive" enemies and "Static" enemies.

Fixed an issue where the "Alive" enemies "giveaway" (an animation wherein their heads move back and fourth slowly) didn't loop. This was balanced, however, with the first feature mentioned.