Changes were made to the Rocket Launcher weapon in order to ensure that animations are displayed correctly, i.e. a reload animation was created specifically for the Rocket Launcher and its idle animation now appears in the correct location.
Changes were made to the scripts utilizing Photon's RPC's to ensure that, when in single player, normal method calls are made. This is a work around to a bug wherein, when a new level was loaded within single player mode, damage could not be dealt to the player or the enemies; RPC's were not firing correctly.