Added rudimentary jumping for enemies: there is currently no animation and the enemies instantly transition from jump to land. Additionally, there is currently a bug wherein enemies will not be able to chase the player if the go directly from one elevated service to another: the enemy will attempt to get down from the wrong landing spot due to the way proximity to player determines which spot is chosen (the closest to the player).
Added simple textures to the Box map, specifically the floor and column.

Made fundamental changes to the Machinegun in order to make the model more workable. However, this has caused something of a reversion regarding the weapons reload animation.

Added a shooting animation for the Machinegun, though it is not in a noticeable state.

Added a (WIP) HUD background for the player: currently only functional on the facing worlds map.

Made changes to the initial instantiation of the players weapon towards making the system more dynamic. Currently a weapon needs to be manually placed on the player before weapons are changed, though with these (as well as necessary future changes) the weapon that the player starts with will be much more easily editable. However, changes made toward this have created a bug wherein the player's initial weapon is instantiated quite far away from them, meaning that no model can be seen (although the player can still shoot as expected); changing weapons will bring the weapon models back into the players view.