Fixed an issue wherein the rocket deployed from the rocket launcher would detonate on the player's collider.
Added weapons for the "outside" of the player: now, in multiplayer, other players are able to see which weapon the other player has selected. With this, a default transition between the weapons is displayed and a rudimentary reload animation is played.

In order to achieve this, small changes were made to the code in order to ensure that particular animations were played at particular intervals in the game, i.e. setting the correct integer value when a weapon was changed and setting the reload bool to true an false in parity with the animation played from the first person perspective. On the animation and modeling side of things (Blender), animations were created using the "outside" player model and already created weapons; each weapon needed a simple idle and reload animation. Regarding the utilization of Unity's animation utility, the created blender file was somewhat altered to ensure that it could be "humanoid" and weapons could be animated; this was done using animation masking.