Added scoring system as new UI element; TAB can be pressed to activate and deactivate UI associated with this. Time and the amount of static enemies killed are factored in this.
Changes were made to the model and animations for the shotgun, albeit small, making the model and associated animations more visually pleasing.

3D sounds for the enemies to make when they are alerted or "awake"  were added: a random selection is made from the two available MP3's.

Changes were made regarding the way in which enemies chase the player: now they are able to execute the attack animation (which involves the upper body) regardless of whether they are walking or standing, both states often utilized. With this, players are no longer to as easily escape the enemies and are certain to receive some damage before evading the enemy's chase range.

The damage caused by the rocket (from the rocket launcher) was re-balanced in order to incorporate the enemies using 6 hit boxes rather than one.