Updated project to Unity 5. This changed various files for optimization meaning that they are now incompatible with earlier version of Unity.
Made changes to code to ensure that the cursor still locked and became visible appropriately: adjustments were made to the API in Unity 5.

Made changes to the UI manager class to remove bug that made the application crash when exiting. Applicaion.Quit() is not as functional as thought, System.Diagnostics.Process.GetCurrentProcess().Kill() seems to do the job, however.