Added functionality for weapons creating sounds when they are fired, though currently a sound has only been created for the machine gun (and is being used across all weapons).
Added functionality for alive enemies wherein they are able to chase the player if they have received any damage once alives have been alerted.

changed the way in which the player's score is generated to deduct points for the killing of alive enemies.

Fixed bug wherein the game would end immediately once the player selects single player, in the case that they have already played a game (are in the same instance of the game): the amount of static enemies was not being reset when the game ended.

Made changes to the navigation mesh of the level to ensure that enemies are able to properly get to areas that are supposed to be available to them.