Made changes within both the input manager and the ui manager to make it so that the cursor turns invisible when the user is to enter a new control-more intuitive
Updated the way in which values are passed to the text of the control ui; it now will update everything when a control is changed. less efficient, but is a simple get string set string. could always opt to go back to updating individually and only update everything on special occasions

Listening for the input is still a bit off, user sometimes needs to press multiple times for their request to be heard, but by locking the fields and making the mouse disappear until a key has been submitted, users should have less trouble.

Added to the axis class to that if the individually set axis are the same (both the positive and negative side of an instance of my axis class have the same axis string) then Unity's axis is returned rather than the complex "turn key or axis side into axis" sequence. The latter needs optimizing.